removed unused files and reorganized game logic files
