+++ gameLogic/cardGenerator.js
