+++ gameLogic/checkTrading.js
