+++ gameLogic/checkWonderLevelOneTrading.js
