+++ gameLogic/checkWonderLevelThreeTrading.js
