+++ gameLogic/checkWonderLevelTwoTrading.js
