+++ gameLogic/militaryComparison.js
