+++ gameLogic/tradeModalHelperFuncs.js
