--- script/seed.spec.js
@@ -1,8 +1,0 @@
-'use strict'
-/* global describe beforeEach it */
-
-const seed = require('./seed')
-
-describe('seed script', () => {
-  it('completes successfully', seed)
-})
